mediator: massive refactoring of mediator
In this commit I shuffled around almost all parts of the mediator, and
most importantly, mediator now runs in a separate process. All this fuss
was due to the this deprecation warning:
https://werkzeug.palletsprojects.com/en/2.0.x/serving/#shutting-down-the-server

So I figured I should move to a new shutdown method. Well, it complicated
everything for no solid reason, if you ask me. But here we go, now we are doing
things "the right way".