Correct command of Installation via pipx
There was a little mistake in the command that was written to install brotab using pipx. 
It was written `pipx brotab` but the command that works is `pipx install brotab`.
Also, in my ubuntu desktop, pipx wasn't installed, so I need to install it via the command 'sudo apt install pipx'. 
This worked for me, so I thought to correct this mistake in the README to avoid any confusion in people using it in future.